requirements.txt
pytest~=7.4.3
fastapi~=0.104.1
requests~=2.31.0
validators~=0.22.0
torch~=2.1.0
matplotlib~=3.8.1
uvicorn~=0.23.2
Pillow~=10.1.0
transformers~=4.35.0
httpx~=0.25.0
aiofiles==23.1.0
aiogram==3.0.0b7
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
certifi==2023.5.7
charset-normalizer==3.1.0
frozenlist==1.3.3
idna==3.4
magic-filter==1.0.9
multidict==6.0.4
pydantic==1.10.11
pytz==2023.3
requests==2.31.0
typing_extensions==4.7.1
urllib3==2.0.3
yarl==1.9.2